古战场逃兵
修复了位于无月之影控制台的释放NPC功能；
增加了雷达可显示的NPC类别；
修改了数据库中type字段的长度；

+++ gamedata/sql/all.sql
@@ -186,7 +186,7 @@ CREATE TABLE IF NOT EXISTS `acbra2_newsinfo` (
 DROP TABLE IF EXISTS `acbra2_players`;
 CREATE TABLE IF NOT EXISTS `acbra2_players` (
   `pid` mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
-  `type` tinyint(3) unsigned NOT NULL DEFAULT '0',
+  `type` smallint(5) unsigned NOT NULL DEFAULT '0',
   `name` char(15) NOT NULL DEFAULT '',
   `pass` char(32) NOT NULL DEFAULT '',
   `ip` char(15) NOT NULL DEFAULT '',
